Fix up ViewPostUI issues

+++ src/pageturners/ui/modules/ViewPostUI.java
@@ -2,22 +2,16 @@ package pageturners.ui.modules;
 
 import java.text.DateFormat;
 
-import javafx.scene.paint.Color;
-import javafx.geometry.Insets;
 import javafx.geometry.Orientation;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
-import javafx.scene.layout.CornerRadii;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Font;
-import jdk.management.resource.internal.TotalResourceContext;
 import pageturners.controls.ControlDirectory;
 import pageturners.controls.LoginControl;
 import pageturners.controls.ViewPostControl;
@@ -84,15 +78,13 @@ public class ViewPostUI extends UIElement {
 
         if (deleteButton != null) {
             deleteButton.minWidthProperty().bind(deleteButton.prefWidthProperty().add(100));
+            deleteButton.maxWidthProperty().bind(deleteButton.prefWidthProperty().add(100));
         }
 
         Label contents = new Label(postContent.userPost.isDeleted ? "This post was deleted." : postContent.userPost.contents);
+        contents.setFont(Font.font(18));
         contents.setWrapText(true);
         contents.setPrefWidth(Double.MAX_VALUE);
-
-        Label commentsHeader = new Label("Comments:");
-        commentsHeader.setFont(Font.font(15));
-
         int totalCols = 4;
 
         layout.add(title, 0, 0, totalCols, 1);
@@ -106,13 +98,10 @@ public class ViewPostUI extends UIElement {
         layout.add(new Separator(Orientation.HORIZONTAL), 0, 3, totalCols, 1);
         layout.add(contents, 0, 4, totalCols, 1);
         layout.add(new Separator(Orientation.HORIZONTAL), 0, 5, totalCols, 1);
-        layout.add(commentsHeader, 0, 6, totalCols, 1);
 
-        if (deleteButton != null) {
-            GridPane.setHgrow(spacer, Priority.ALWAYS);
-        }
+        GridPane.setHgrow(spacer, Priority.ALWAYS);
 
-        int commentRow = 7;
+        int commentRow = 6;
 
         if (postContent.userComments.size() == 0) {
             Label noCommentsLabel = new Label("There are no comments at the moment. Be the first!");
@@ -121,7 +110,8 @@ public class ViewPostUI extends UIElement {
         else {
             for (UserCommentObject comment : postContent.userComments) {
                 Region commentUI = getCommentUI(postContent, comment);
-                layout.add(commentUI, 0, commentRow++, 3, 1);
+                layout.add(commentUI, 0, commentRow++, totalCols, 1);
+                GridPane.setHgrow(commentUI, Priority.ALWAYS);
             }
         }
 
@@ -133,42 +123,49 @@ public class ViewPostUI extends UIElement {
         else {
             AddUserCommentUI addUserCommentUI = new AddUserCommentUI(controlDirectory, mainWindowBodyUI, postContent.userPost);
             layout.add(addUserCommentUI.getNode(), 0, commentRow++, totalCols, 1);
+            addUserCommentUI.getNode().setMinHeight(50);
         }
 
         show(layout);
     }
 
     private Region getCommentUI(UserPostContentObject postContent, UserCommentObject comment) {
-        GridPane layout = new GridPane();
+        VBox vLayout = new VBox();
+        HBox hLayout = new HBox();
 
-        Label author = new Label(comment.user.username + ":");
+        Label author = new Label(comment.user.username + "");
         Label commentDate = new Label(DATE_FORMAT.format(comment.commentDate));
+        Region spacer = new Region();
 
-        int totalCols = 0;
+        hLayout.getChildren().addAll(author, new Separator(Orientation.VERTICAL), commentDate, spacer);
 
-        layout.add(author, totalCols++, 0);
-        layout.add(commentDate, totalCols++, 0);
+        HBox.setHgrow(spacer, Priority.ALWAYS);
 
-        if (loginControl.getAdminObject() != null) {
-            DeleteCommentUI deleteCommentUI = new DeleteCommentUI(controlDirectory, mainWindowBodyUI, postContent.userPost, comment);
-
-            layout.add(deleteCommentUI.getNode(), totalCols++, 0);
+        if (!comment.isDeleted) {
+            if (loginControl.getAdminObject() != null) {
+                DeleteCommentUI deleteCommentUI = new DeleteCommentUI(controlDirectory, mainWindowBodyUI, postContent.userPost, comment);
+    
+                hLayout.getChildren().add(deleteCommentUI.getNode());
+            }
+            else if (loginControl.getUserObject() != null && loginControl.getUserObject().id == comment.userId) {
+                DeleteUserCommentUI deleteUserCommentUI = new DeleteUserCommentUI(controlDirectory, mainWindowBodyUI, postContent.userPost, comment);
+    
+                hLayout.getChildren().add(deleteUserCommentUI.getNode());
+            }
         }
-        else if (loginControl.getUserObject() != null && loginControl.getUserObject().id == postContent.author.id) {
-            DeleteUserCommentUI deleteUserCommentUI = new DeleteUserCommentUI(controlDirectory, mainWindowBodyUI, postContent.userPost, comment);
 
-            layout.add(deleteUserCommentUI.getNode(), totalCols++, 0);
-        }
-
-        Label commentContents = new Label(comment.contents);
+        Label commentContents = new Label(comment.isDeleted ? "This comment has been deleted." : comment.contents);
+        commentContents.setFont(Font.font(15));
         commentContents.setWrapText(true);
 
-        layout.add(commentContents, 0, 1, totalCols, 1);
+        Separator separator = new Separator(Orientation.HORIZONTAL);
 
-        layout.setHgap(10);
-        layout.setVgap(5);
+        vLayout.getChildren().addAll(hLayout, commentContents, separator);
 
-        return layout;
+        hLayout.setSpacing(15);
+        vLayout.setSpacing(5);
+
+        return vLayout;
     }
 
     private void displayFailure() {
